Query for getting signed out drivers and details
Created to queries, one of retreiving the names of drivers that have
singed out. It returns only drivers signed out that day and where time
back in is null.

Created a query for retreiving all the details required to sign back in.
It returns the duty, pda numbers and van number they signed out and
fills in the fields on the form based on these values.