Queryfor all van details and removing a van
Created a query to return the all the attributes for the vans in the
system. The values can then be used as required. Currently used in
removing a van and will fill in the fields for serial number and vehicle
number based on reg number.

Created a query for removing a van from those available. The van is not
removed entirely but its attributes for being available to be signed out
are set to 0.